break apart responsibilities
